+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+- Fix Bing image provider
 - Fix bug where if we did not properly load Bing image it broke app state saving.
 - Update dependencies